functions/requirements.txt
firebase_functions~=0.1.0
google-genai~=1.8.0
requests~=2.28.1